+++ src/popup.js
@@ -16,400 +16,20 @@
 *                                                                                 *
 * Home: https://github.com/Crivella/ocr_extension                                 *
 **********************************************************************************/
-import React, { createContext, useCallback, useContext, useEffect, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import ReactDOM from 'react-dom/client';
 
-import { QueryClient, QueryClientProvider, useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { QueryClient, QueryClientProvider, useQuery } from '@tanstack/react-query';
 
-import { handshake, post } from "./utils/API";
+import { GlobalContext } from "./components/context";
+import { EndpointField, FontScaleField, RGBField } from "./components/fields";
+import { LangUnit } from "./components/submitUnitLang";
+import { ModelUnit } from "./components/submitUnitModel";
+import { ThemeSwitch } from "./components/themeSwitch";
+import { handshake } from "./utils/API";
 
-const MaxLangField = 25;
-const GlobalContext = createContext();
 const queryClient = new QueryClient();
 
-/*
-React component to switch the theme of the popup.
-*/
-function ThemeSwitch() {
-    const [theme, setTheme] = useState('dark');
-
-    useEffect(() => {
-        browser.storage.local.get().then((res) => {
-            setTheme(res.popupTheme || 'dark');
-        })
-    }, [])
-
-    useEffect(() => {
-        document.documentElement.style.setProperty('--theme-body-bg-color', `var(--theme-${theme}-body-bg-color)`);
-        document.documentElement.style.setProperty('--theme-color', `var(--theme-${theme}-color)`);
-        document.documentElement.style.setProperty('--theme-text-bg-color', `var(--theme-${theme}-text-bg-color)`);
-    }, [theme])
-
-    const onClick = (e) => {
-        e.preventDefault();
-        e.stopPropagation();
-        const newTheme = theme === 'light' ? 'dark' : 'light';
-        browser.storage.local.set({popupTheme: newTheme});
-        setTheme(newTheme);
-    }
-
-    return (
-        <div onClick={onClick}>
-            <label class="theme-switch">
-                <input type="checkbox" checked={theme === 'light'} />
-                <span class="slider round"></span>
-            </label>
-        </div>
-    )
-}
-
-/*
-React component to draw a field for the endpoint.
-The field will be highlighted in green or red depending on the success of the handshake.
-*/
-function EndpointField() {
-    const { endpoint, setEndpoint, successEndpoint: success } = useContext(GlobalContext);
-
-    const [endpointInput, setEndpointInput] = useState(endpoint);
-    const [myClass, setClass] = useState(''); // ['success', 'error', '']
-
-    const onSubmit = (e) => {
-        e.preventDefault();
-        console.log(endpointInput);
-        setEndpoint(endpointInput);
-    }
-
-    useEffect(() => {
-        setEndpointInput(endpoint);
-    }, [endpoint])
-
-    useEffect(() => {
-        if (success === null) {
-            setClass('');
-        } else if (success) {
-            setClass('success');
-        } else {
-            setClass('error');
-        }
-    }, [success])
-
-    
-    return (
-        <div className="field">
-            <label htmlFor="endpoint">Endpoint</label>
-            <input 
-                type="text" id="endpoint" name="endpoint" 
-                className={myClass}
-                value={endpointInput} 
-                onChange={(e) => setEndpointInput(e.target.value)}
-                onKeyDown={(e) => {
-                    if (e.key === 'Enter') {
-                        onSubmit(e);
-                    }}
-                } 
-                />
-            <input type="submit" value="Apply" onClick={onSubmit} />
-        </div>
-    )
-}
-
-/*
-React component to draw a field for the font scale.
-*/
-function FontScaleField() {
-    const { fontScale, setFontScale } = useContext(GlobalContext);
-
-    return (
-        <div className="field">
-            <label htmlFor="font-scale">Font Scale</label>
-            <input
-                type="number" id="font-scale" name="font-scale"
-                value={fontScale}
-                min="0.1" max="3" step="0.1"
-                onChange={(e) => setFontScale(e.target.value)}
-                onScroll={(e) => console.log(e)}
-            />
-        </div>
-    )
-}
-
-/*
-React component to draw a field/s for the font RGB color.
-*/
-function RGBField() {
-    const { RGB, setRGB } = useContext(GlobalContext);
-    const [hexRGB, setHexRGB] = useState();
-
-    useEffect(() => {
-        if (RGB[0] !== undefined) {
-            setHexRGB(`#${RGB.map((e) => e.toString(16).padStart(2, '0')).join('')}`);
-        }
-    }, [RGB])
-
-    // https://stackoverflow.com/questions/61821924/firefox-addon-input-type-color-closing-addon-popup-on-firefox
-    // const onPick = (e) => {
-    //     console.log(e.target.value);
-    //     const hex = e.target.value.slice(1);
-    //     const r = parseInt(hex.slice(0, 2), 16);
-    //     const g = parseInt(hex.slice(2, 4), 16);
-    //     const b = parseInt(hex.slice(4, 6), 16);
-    //     setRGB([r, g, b]);
-    //     // setRGB(e.target.value);
-    // }
-
-    return (
-        <div className="field">
-            <label htmlFor="R">R</label>
-            <input
-                type="number" id="R" name="R"
-                value={RGB[0]}
-                min="0" max="255" step="1"
-                onChange={(e) => setRGB([e.target.value, RGB[1], RGB[2]])}
-            />
-            <label htmlFor="G">G</label>
-            <input
-                type="number" id="G" name="G"
-                value={RGB[1]}
-                min="0" max="255" step="1"
-                onChange={(e) => setRGB([RGB[0], e.target.value, RGB[2]])}
-            />
-            <label htmlFor="B">B</label>
-            <input
-                type="number" id="B" name="B"
-                value={RGB[2]}
-                min="0" max="255" step="1"
-                onChange={(e) => setRGB([RGB[0], RGB[1], e.target.value])}
-            />
-            {/* <input type="color" id="color" name="color" value={hexRGB} onChange={onPick} /> */}
-        </div>
-    )
-}
-
-
-/*
-Generic React component to draw a field for a list of selections.
-The field will be highlighted in green or red depending on the success variable.
-EG. field included in a submission form, that will set success for all its child components.
-*/
-function SelectField({label, name, options, option_names, value, setValue, success}) {
-    const [myClass, setClass] = useState(''); // ['success', 'error', '']
-
-    const onChange = (e) => {
-        setValue(e.target.value);
-    }
-
-    useEffect(() => {
-        if (success === null) {
-            setClass('');
-        } else if (success) {
-            setClass('success');
-        } else {
-            setClass('error');
-        }
-    }, [success])
-
-    return (
-        <div className="field">
-            <label htmlFor={name}>{label}</label>
-            <select className={myClass} name={name} id={name} value={value} onChange={onChange}>
-                <option value="">--- SELECT ---</option> 
-                {options.map((option, idx) => (
-                    <option key={idx} value={option}>{option_names[idx]}</option>
-                ))}
-            </select>
-        </div>
-    )
-}
-
-/*
-Field for the list of available box models.
-*/
-function BOXModelSelect({ success = null }) {
-    const { boxModels, boxModel, setBoxModel } = useContext(GlobalContext);
-
-    return <SelectField 
-        label="BOX Model" 
-        name="box-model" 
-        options={boxModels}
-        option_names={boxModels}
-        value={boxModel} 
-        setValue={setBoxModel} 
-        success={success} />
-}
-
-/*
-Field for the list of available OCR models.
-*/
-function OCRModelSelect({ success = null }) {
-    const { ocrModels, ocrModel, setOcrModel } = useContext(GlobalContext);
-
-    return <SelectField 
-        label="OCR Model" 
-        name="ocr-model" 
-        options={ocrModels} 
-        option_names={ocrModels} 
-        value={ocrModel} 
-        setValue={setOcrModel} 
-        success={success} />
-}
-
-/*
-Field for the list of available TSL models.
-*/
-function TSLModelSelect({ success = null }) {
-    const { tslModels, tslModel, setTslModel } = useContext(GlobalContext);
-
-    return <SelectField
-        label="Translation Model"
-        name="tsl-model"
-        options={tslModels}
-        option_names={tslModels}
-        value={tslModel}
-        setValue={setTslModel}
-        success={success} />
-}
-
-/*
-Field for the list of available source languages.
-*/
-function LanguageSrcSelect({ success = null }) {
-    const { langChoices, langSrc, setLangSrc, langChoicesHR } = useContext(GlobalContext);
-
-    return <SelectField
-        label="Source Language"
-        name="lang-src"
-        options={langChoices}
-        option_names={langChoicesHR.map((e) => e.length > MaxLangField ? e.slice(0, MaxLangField-3) + '...' : e)}
-        value={langSrc}
-        setValue={setLangSrc}
-        success={success} />
-}
-
-/*
-Field for the list of available destination languages.
-*/
-function LanguageDstSelect({ success = null }) {
-    const { langChoices, langDst, setLangDst, langChoicesHR } = useContext(GlobalContext);
-    
-    return <SelectField
-        label="Destination Language"
-        name="lang-dst"
-        options={langChoices}
-        option_names={langChoicesHR.map((e) => e.length > MaxLangField ? e.slice(0, MaxLangField-3) + '...' : e)}
-        value={langDst}
-        setValue={setLangDst}
-        success={success} />
-}
-
-/*
-Generic React component to draw a submission form.
-The props `target` and `data` should be given by the parent component.
-This abstract the creation of a form, and handling data submission and success/failure.
-*/
-function SubmitUnit({children, target, data}) {
-    const queryClient = useQueryClient();
-
-    const { endpoint } = useContext(GlobalContext);
-
-    const [success, setSuccess] = useState(null);
-
-    const updateMutation = useMutation({
-        mutationFn: (data) => post(endpoint, target, data), 
-        onError: () => {
-            console.log('error');
-            setSuccess(false);
-        },
-        onSuccess: () => {
-            console.log('success');
-            setSuccess(true);
-        },
-        onSettled: () => {
-            queryClient.invalidateQueries(['endpoint']);
-        }
-    });
-
-    const onSubmit = useCallback(async (e) => {
-        e.preventDefault();
-        console.log(updateMutation.status)
-        if (updateMutation.isLoading) {
-            return;
-        }
-        console.log('submitting', data);
-        updateMutation.mutate(data);
-    }, [updateMutation, data]);
-
-    useEffect(() => {
-        setSuccess(null);
-    }, [data])
-
-    const newProps = {
-        success: success,
-    }
-
-    return (
-        <div className="subunit">
-            {children.map((e) => {
-                // return e;
-                let newChild = {...e};
-                if (typeof(e.type) === 'function') newChild.props = {...newProps, ...e.props};
-                return newChild;
-                }
-            )}
-            {
-                updateMutation.isLoading ? 
-                <div className="loading">Loading...</div> 
-                : 
-                <button onClick={onSubmit}>Submit</button>
-            }
-        </div>
-    )
-}
-
-/*
-React component to draw the form for the model selection.
-*/
-function ModelUnit() {
-    const [data, setData] = useState({}); 
-    const { boxModel, ocrModel, tslModel } = useContext(GlobalContext);
-    
-    useEffect(() => {
-        setData({
-            box_model_id: boxModel,
-            ocr_model_id: ocrModel,
-            tsl_model_id: tslModel,
-        })
-    }, [boxModel, ocrModel, tslModel])
-
-    return (
-        <SubmitUnit target="set_models" data={data}>
-            <BOXModelSelect />
-            <OCRModelSelect />
-            <TSLModelSelect />
-        </SubmitUnit>   
-    )
-}
-
-/*
-React component to draw the form for the language selection.
-*/
-function LangUnit() {
-    const [data, setData] = useState({});
-    const { langSrc, langDst } = useContext(GlobalContext);
-
-    useEffect(() => {
-        setData({
-            lang_src: langSrc,
-            lang_dst: langDst,
-        })
-    }, [langSrc, langDst])
-
-    return (
-        <SubmitUnit target="set_lang" data={data}>
-            <LanguageSrcSelect />
-            <LanguageDstSelect />
-        </SubmitUnit>
-    )
-}
 
 function Checkbox({label, name, value, setValue}) {
     return (
@@ -461,10 +81,6 @@ function PopUp() {
     });
 
     useEffect(() => {
-        console.log('ENDPOINT', endpoint);
-    }, [endpoint])
-
-    useEffect(() => {
         console.log('QUERY', query);
         if (query.data) {
             setBoxModels(query.data.BOXModels || []);
@@ -510,7 +126,7 @@ function PopUp() {
 /*
 React component to handle the context and messaging with the background script.
 */
-function Hub() {
+export function Hub() {
     const [fontScale, setFontScale] = useState();
     const [RGB, setRGB] = useState([undefined, undefined, undefined]);
     const [langSrc, setLangSrc] = useState('');
